refactor: merge pg-format package into pg-core
- Move format.ts and reserved.ts from pg-format to pg-core/src/
- Move format tests to pg-core/src/__tests__/
- Update all imports from @graphile/pg-format to @graphile/pg-core
- Remove pg-format package entirely
- Export formatting utilities from pg-core index
