+++ packages/pg-adapter-pglite/src/index.ts
@@ -69,7 +69,11 @@ function createPGLiteConnection(db: PGlite): PgConnection {
   }
 
   return {
-    async execute(sql: string, params?: any[], options?: { prepare?: boolean }): Promise<void> {
+    async execute(
+      sql: string,
+      params?: any[],
+      options?: { prepare?: boolean },
+    ): Promise<void> {
       try {
         await db.query(sql, params);
       } catch (error) {
@@ -80,7 +84,7 @@ function createPGLiteConnection(db: PGlite): PgConnection {
     async query<T extends MaybeRow = any>(
       sql: string,
       params?: any[],
-      options?: { prepare?: boolean }
+      options?: { prepare?: boolean },
     ): Promise<PgQueryResult<T>> {
       try {
         const result = await db.query<T>(sql, params);
@@ -158,7 +162,11 @@ function createPGLiteClient(
   wrapError: (error: unknown) => PgAdapterError,
 ): PgClient {
   return {
-    async execute(sql: string, params?: any[], options?: { prepare?: boolean }): Promise<void> {
+    async execute(
+      sql: string,
+      params?: any[],
+      options?: { prepare?: boolean },
+    ): Promise<void> {
       try {
         await db.query(sql, params);
       } catch (error) {
@@ -169,7 +177,7 @@ function createPGLiteClient(
     async query<T extends MaybeRow = any>(
       sql: string,
       params?: any[],
-      options?: { prepare?: boolean }
+      options?: { prepare?: boolean },
     ): Promise<PgQueryResult<T>> {
       try {
         const result = await db.query<T>(sql, params);
@@ -194,7 +202,11 @@ function createPGLiteTransactionClient(
   wrapError: (error: unknown) => PgAdapterError,
 ): PgClient {
   return {
-    async execute(sql: string, params?: any[], options?: { prepare?: boolean }): Promise<void> {
+    async execute(
+      sql: string,
+      params?: any[],
+      options?: { prepare?: boolean },
+    ): Promise<void> {
       try {
         await tx.query(sql, params);
       } catch (error) {
@@ -205,7 +217,7 @@ function createPGLiteTransactionClient(
     async query<T extends MaybeRow = any>(
       sql: string,
       params?: any[],
-      options?: { prepare?: boolean }
+      options?: { prepare?: boolean },
     ): Promise<PgQueryResult<T>> {
       try {
         const result = await tx.query<T>(sql, params);
