+++ packages/pg-adapter-postgres-js/package.json
@@ -53,7 +53,6 @@
     "postgres": "^3.0.0"
   },
   "devDependencies": {
-    "@graphile/pg-format": "workspace:^",
     "@types/node": "^22.16.5",
     "postgres": "^3.4.7",
     "tsdown": "^0.12.9",
