+++ packages/pg-adapter-postgres-js/src/index.ts
@@ -76,10 +76,16 @@ function createPostgresJsConnectionInternal(
   }
 
   return {
-    async execute(sqlText: string, params?: any[], options?: { prepare?: boolean }): Promise<void> {
+    async execute(
+      sqlText: string,
+      params?: any[],
+      options?: { prepare?: boolean },
+    ): Promise<void> {
       try {
         if (params) {
-          await sql.unsafe(sqlText, params, { prepare: options?.prepare ?? false });
+          await sql.unsafe(sqlText, params, {
+            prepare: options?.prepare ?? false,
+          });
         } else {
           await sql.unsafe(sqlText, [], { prepare: options?.prepare ?? false });
         }
@@ -91,12 +97,16 @@ function createPostgresJsConnectionInternal(
     async query<T extends MaybeRow = any>(
       sqlText: string,
       params?: any[],
-      options?: { prepare?: boolean }
+      options?: { prepare?: boolean },
     ): Promise<PgQueryResult<T>> {
       try {
         const result = params
-          ? await sql.unsafe<T[]>(sqlText, params, { prepare: options?.prepare ?? false })
-          : await sql.unsafe<T[]>(sqlText, [], { prepare: options?.prepare ?? false });
+          ? await sql.unsafe<T[]>(sqlText, params, {
+              prepare: options?.prepare ?? false,
+            })
+          : await sql.unsafe<T[]>(sqlText, [], {
+              prepare: options?.prepare ?? false,
+            });
 
         // postgres.js attaches metadata to the array
         const pgResult = result as any;
@@ -170,12 +180,20 @@ function createPostgresJsConnectionInternal(
 
 function createPostgresJsClient(reserved: ReservedSql): PgClient {
   return {
-    async execute(sql: string, params?: any[], options?: { prepare?: boolean }): Promise<void> {
+    async execute(
+      sql: string,
+      params?: any[],
+      options?: { prepare?: boolean },
+    ): Promise<void> {
       try {
         if (params) {
-          await reserved.unsafe(sql, params, { prepare: options?.prepare ?? false });
+          await reserved.unsafe(sql, params, {
+            prepare: options?.prepare ?? false,
+          });
         } else {
-          await reserved.unsafe(sql, [], { prepare: options?.prepare ?? false });
+          await reserved.unsafe(sql, [], {
+            prepare: options?.prepare ?? false,
+          });
         }
       } catch (error) {
         throw wrapError(error);
@@ -185,12 +203,16 @@ function createPostgresJsClient(reserved: ReservedSql): PgClient {
     async query<T extends MaybeRow = any>(
       sql: string,
       params?: any[],
-      options?: { prepare?: boolean }
+      options?: { prepare?: boolean },
     ): Promise<PgQueryResult<T>> {
       try {
         const result = params
-          ? await reserved.unsafe<T[]>(sql, params, { prepare: options?.prepare ?? false })
-          : await reserved.unsafe<T[]>(sql, [], { prepare: options?.prepare ?? false });
+          ? await reserved.unsafe<T[]>(sql, params, {
+              prepare: options?.prepare ?? false,
+            })
+          : await reserved.unsafe<T[]>(sql, [], {
+              prepare: options?.prepare ?? false,
+            });
 
         const pgResult = result as any;
         return {
@@ -207,10 +229,16 @@ function createPostgresJsClient(reserved: ReservedSql): PgClient {
 
 function createPostgresJsTransactionClient(txSql: Sql): PgClient {
   return {
-    async execute(sql: string, params?: any[], options?: { prepare?: boolean }): Promise<void> {
+    async execute(
+      sql: string,
+      params?: any[],
+      options?: { prepare?: boolean },
+    ): Promise<void> {
       try {
         if (params) {
-          await txSql.unsafe(sql, params, { prepare: options?.prepare ?? false });
+          await txSql.unsafe(sql, params, {
+            prepare: options?.prepare ?? false,
+          });
         } else {
           await txSql.unsafe(sql, [], { prepare: options?.prepare ?? false });
         }
@@ -222,12 +250,16 @@ function createPostgresJsTransactionClient(txSql: Sql): PgClient {
     async query<T extends MaybeRow = any>(
       sql: string,
       params?: any[],
-      options?: { prepare?: boolean }
+      options?: { prepare?: boolean },
     ): Promise<PgQueryResult<T>> {
       try {
         const result = params
-          ? await txSql.unsafe<T[]>(sql, params, { prepare: options?.prepare ?? false })
-          : await txSql.unsafe<T[]>(sql, [], { prepare: options?.prepare ?? false });
+          ? await txSql.unsafe<T[]>(sql, params, {
+              prepare: options?.prepare ?? false,
+            })
+          : await txSql.unsafe<T[]>(sql, [], {
+              prepare: options?.prepare ?? false,
+            });
 
         const pgResult = result as any;
         return {
