+++ packages/pg-core/README.md
@@ -2,7 +2,7 @@
 
 Core PostgreSQL adapter interfaces and utilities for Graphile tools.
 
-This package provides the foundational interfaces and types that all PostgreSQL adapters implement, enabling a consistent stream-first API across different PostgreSQL client libraries.
+This package provides the foundational interfaces and types that all PostgreSQL adapters implement, enabling a consistent stream-first API across different PostgreSQL client libraries. It also includes SQL formatting utilities for safely building dynamic SQL queries.
 
 ## Installation
 
@@ -35,6 +35,24 @@ for await (const batch of result.batches(100)) {
 
 // Get count without loading all rows
 const count = await result.count();
+```
+
+## SQL Formatting Utilities
+
+The package includes SQL formatting functions to safely build dynamic SQL queries:
+
+```typescript
+import { format, ident, literal } from "@graphile/pg-core";
+
+// Format with placeholders
+const query = format("SELECT * FROM %I WHERE id = %L", "users", 123);
+// Result: SELECT * FROM users WHERE id = '123'
+
+// Quote identifiers
+const table = ident("my-table"); // Result: "my-table"
+
+// Quote literals
+const value = literal("O'Brien"); // Result: 'O''Brien'
 ```
 
 ## Usage
