+++ packages/pg-core/src/index.ts
@@ -39,7 +39,11 @@ export interface PgClient {
   /**
    * Execute a query without returning results (for DDL, DML, transaction control)
    */
-  execute(sql: string, params?: any[], options?: { prepare?: boolean }): Promise<void>;
+  execute(
+    sql: string,
+    params?: any[],
+    options?: { prepare?: boolean },
+  ): Promise<void>;
 
   /**
    * Execute a query and return results with metadata
@@ -47,7 +51,7 @@ export interface PgClient {
   query<T extends MaybeRow = Row>(
     sql: string,
     params?: any[],
-    options?: { prepare?: boolean }
+    options?: { prepare?: boolean },
   ): Promise<PgQueryResult<T>>;
 }
 
@@ -178,3 +182,4 @@ export function createPgPool(adapter: PgConnection): PgPool {
 }
 
 export { createPgHelpers } from "./helpers";
+export * from "./format";
