--- packages/pg-format/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "declarationDir": "./dist",
-    "outDir": "./dist",
-    "moduleResolution": "bundler",
-    "isolatedDeclarations": true
-  },
-  "include": ["src"]
-}
